requirements.txt
# Core Dependencies
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0,<2.0.0
Pillow>=10.0.0

# Deep Learning (will auto-install compatible versions)
torch>=2.0.0
torchvision>=0.15.0

# Backend API
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0

# Communication
paho-mqtt>=1.6.0
requests>=2.31.0

# Utilities
python-dotenv>=1.0.0
PyYAML>=6.0.0
python-dateutil>=2.8.0

# Video Processing
imutils>=0.5.4

# Monitoring and Logging
loguru>=0.7.0

# Development
pytest>=7.4.0
